requirements.txt
pdf2image==1.16.3
Pillow==9.5.0
PyPDF2==3.0.1
streamlit==1.24.1
poppler-utils==0.1.0